Assignment-2/folder/requirements.txt
Flask==3.0.0
Flask_APScheduler==1.13.1
Requests==2.31.0
aiohttp==3.9.3
aiosignal==1.3.1
uvicorn==0.27.1
fastapi==0.110.0
mysql.connector
aiodocker==0.21.0

